refactor(macos): split terminal host runtime owners
