+++ clients/apple/alan-macos/Services/Terminal/TerminalBootResolution.swift
@@ -1,4 +1,3 @@
-import CoreGraphics
 import Foundation
 
 #if os(macOS)
@@ -978,575 +977,4 @@ private struct SurfaceRecreationIdentity: Equatable {
     let launchStrategy: String?
 }
 
-enum TerminalRuntimeRenderPriority: Int, Codable, Equatable, CaseIterable, Comparable {
-    case hiddenBackground = 0
-    case visibleBackground = 1
-    case foregroundInteractive = 2
-
-    static func < (
-        lhs: TerminalRuntimeRenderPriority,
-        rhs: TerminalRuntimeRenderPriority
-    ) -> Bool {
-        lhs.rawValue < rhs.rawValue
-    }
-
-    var isVisible: Bool {
-        self != .hiddenBackground
-    }
-
-    var isForegroundInteractive: Bool {
-        self == .foregroundInteractive
-    }
-
-    var diagnosticsValue: String {
-        switch self {
-        case .foregroundInteractive:
-            return "foregroundInteractive"
-        case .visibleBackground:
-            return "visibleBackground"
-        case .hiddenBackground:
-            return "hiddenBackground"
-        }
-    }
-
-    var diagnosticsVisibility: String {
-        isVisible ? "visible" : "hidden"
-    }
-}
-
-enum TerminalRenderRefreshReason: String, Equatable {
-    case automatic
-    case catchUp = "catch_up"
-}
-
-struct TerminalRenderCoordinatorMetrics: Codable, Equatable {
-    var wakeupRequests = 0
-    var appTicks = 0
-    var surfaceRefreshes = 0
-    var coalescedSurfaceRefreshes = 0
-    var catchUpRefreshes = 0
-    var cancelledDrains = 0
-    var drainBatches = 0
-    var lastDrainBatchSize = 0
-    var maxDrainBatchSize = 0
-    var lastDrainLatencyMs = 0.0
-    var maxDrainLatencyMs = 0.0
-    var foregroundInteractiveDrains = 0
-    var visibleBackgroundDrains = 0
-    var hiddenBackgroundDrains = 0
-
-    mutating func recordDrain(priority: TerminalRuntimeRenderPriority) {
-        switch priority {
-        case .foregroundInteractive:
-            foregroundInteractiveDrains += 1
-        case .visibleBackground:
-            visibleBackgroundDrains += 1
-        case .hiddenBackground:
-            hiddenBackgroundDrains += 1
-        }
-    }
-
-    mutating func recordDrainBatch(size: Int, latencyMs: Double) {
-        drainBatches += 1
-        lastDrainBatchSize = size
-        maxDrainBatchSize = max(maxDrainBatchSize, size)
-        lastDrainLatencyMs = latencyMs
-        maxDrainLatencyMs = max(maxDrainLatencyMs, latencyMs)
-    }
-}
-
-protocol TerminalRenderCoordinatedHost: AnyObject {
-    var terminalRenderPriority: TerminalRuntimeRenderPriority { get }
-    var isRenderCoordinatorTargetAlive: Bool { get }
-
-    func renderCoordinatorDrainAppTick()
-    func renderCoordinatorRefreshSurface(reason: TerminalRenderRefreshReason)
-}
-
-final class TerminalRenderCoordinator {
-    private struct PendingWakeup {
-        weak var host: TerminalRenderCoordinatedHost?
-        let sequence: Int
-        let enqueuedAt: DispatchTime
-        var requiresSurfaceRefresh: Bool
-        var reason: TerminalRenderRefreshReason
-    }
-
-    private let lock = NSLock()
-    private let automaticallyDrains: Bool
-    private let diagnosticsRecorder: AlanPerformanceDiagnosticsRecorder?
-    private var pendingWakeupsByHost: [ObjectIdentifier: PendingWakeup] = [:]
-    private var drainScheduled = false
-    private var nextSequence = 0
-
-    private var metrics = TerminalRenderCoordinatorMetrics()
-
-    init(
-        automaticallyDrains: Bool = true,
-        diagnosticsRecorder: AlanPerformanceDiagnosticsRecorder? = nil
-    ) {
-        self.automaticallyDrains = automaticallyDrains
-        self.diagnosticsRecorder = diagnosticsRecorder
-    }
-
-    func requestWakeup(
-        from host: TerminalRenderCoordinatedHost,
-        requiresSurfaceRefresh: Bool = true
-    ) {
-        enqueueWakeup(
-            from: host,
-            reason: .automatic,
-            requiresSurfaceRefresh: requiresSurfaceRefresh
-        )
-    }
-
-    func requestCatchUp(from host: TerminalRenderCoordinatedHost) {
-        enqueueWakeup(
-            from: host,
-            reason: .catchUp,
-            requiresSurfaceRefresh: true
-        )
-    }
-
-    func drainPending() {
-        let pendingWakeups = takePendingWakeups()
-        guard !pendingWakeups.isEmpty else { return }
-
-        let drainStartedAt = DispatchTime.now()
-        let maxDrainLatencyMs = pendingWakeups
-            .map { latencyMs(from: $0.enqueuedAt, to: drainStartedAt) }
-            .max() ?? 0
-        updateMetrics { metrics in
-            metrics.recordDrainBatch(size: pendingWakeups.count, latencyMs: maxDrainLatencyMs)
-        }
-
-        for pending in pendingWakeups {
-            guard let host = pending.host,
-                  host.isRenderCoordinatorTargetAlive
-            else {
-                updateMetrics { metrics in
-                    metrics.cancelledDrains += 1
-                }
-                continue
-            }
-
-            let priority = host.terminalRenderPriority
-            updateMetrics { metrics in
-                metrics.appTicks += 1
-                metrics.recordDrain(priority: priority)
-            }
-            let tickStartedAt = performanceDiagnosticsStartTime()
-            host.renderCoordinatorDrainAppTick()
-            if let tickStartedAt {
-                recordDiagnostics(
-                    kind: .ghosttyAppTick,
-                    durationMs: latencyMs(from: tickStartedAt, to: DispatchTime.now()),
-                    priority: priority
-                )
-            }
-
-            guard pending.requiresSurfaceRefresh else { continue }
-            let shouldRefresh = priority.isVisible || pending.reason == .catchUp
-            guard shouldRefresh else {
-                updateMetrics { metrics in
-                    metrics.coalescedSurfaceRefreshes += 1
-                }
-                continue
-            }
-
-            updateMetrics { metrics in
-                metrics.surfaceRefreshes += 1
-                if pending.reason == .catchUp {
-                    metrics.catchUpRefreshes += 1
-                }
-            }
-            let refreshStartedAt = performanceDiagnosticsStartTime()
-            host.renderCoordinatorRefreshSurface(reason: pending.reason)
-            if let refreshStartedAt {
-                recordDiagnostics(
-                    kind: pending.reason == .catchUp
-                        ? .terminalCatchUpRefresh
-                        : .ghosttySurfaceRefresh,
-                    durationMs: latencyMs(from: refreshStartedAt, to: DispatchTime.now()),
-                    priority: priority
-                )
-            }
-        }
-    }
-
-    func metricsSnapshot() -> TerminalRenderCoordinatorMetrics {
-        lock.lock()
-        let snapshot = metrics
-        lock.unlock()
-        return snapshot
-    }
-
-    private func enqueueWakeup(
-        from host: TerminalRenderCoordinatedHost,
-        reason: TerminalRenderRefreshReason,
-        requiresSurfaceRefresh: Bool
-    ) {
-        let shouldScheduleDrain: Bool
-        lock.lock()
-        metrics.wakeupRequests += 1
-        let hostID = ObjectIdentifier(host)
-        if var pending = pendingWakeupsByHost[hostID] {
-            pending.requiresSurfaceRefresh = pending.requiresSurfaceRefresh || requiresSurfaceRefresh
-            if reason == .catchUp {
-                pending.reason = .catchUp
-            }
-            pendingWakeupsByHost[hostID] = pending
-        } else {
-            pendingWakeupsByHost[hostID] = PendingWakeup(
-                host: host,
-                sequence: nextSequence,
-                enqueuedAt: DispatchTime.now(),
-                requiresSurfaceRefresh: requiresSurfaceRefresh,
-                reason: reason
-            )
-            nextSequence += 1
-        }
-        shouldScheduleDrain = automaticallyDrains && !drainScheduled
-        if shouldScheduleDrain {
-            drainScheduled = true
-        }
-        lock.unlock()
-
-        recordDiagnostics(
-            kind: .ghosttyWakeup,
-            durationMs: 0,
-            priority: host.terminalRenderPriority
-        )
-
-        guard shouldScheduleDrain else { return }
-        DispatchQueue.main.async { [weak self] in
-            self?.drainPending()
-        }
-    }
-
-    private func takePendingWakeups() -> [PendingWakeup] {
-        lock.lock()
-        let pending = pendingWakeupsByHost.values.sorted { lhs, rhs in
-            let lhsPriority = lhs.host?.terminalRenderPriority ?? .hiddenBackground
-            let rhsPriority = rhs.host?.terminalRenderPriority ?? .hiddenBackground
-            if lhsPriority != rhsPriority {
-                return lhsPriority > rhsPriority
-            }
-            return lhs.sequence < rhs.sequence
-        }
-        pendingWakeupsByHost.removeAll()
-        drainScheduled = false
-        lock.unlock()
-        return pending
-    }
-
-    private func latencyMs(from start: DispatchTime, to end: DispatchTime) -> Double {
-        let nanos = end.uptimeNanoseconds >= start.uptimeNanoseconds
-            ? end.uptimeNanoseconds - start.uptimeNanoseconds
-            : 0
-        return Double(nanos) / 1_000_000
-    }
-
-    private func performanceDiagnosticsStartTime() -> DispatchTime? {
-        if let diagnosticsRecorder {
-            return diagnosticsRecorder.isEnabled ? DispatchTime.now() : nil
-        }
-        return AlanPerformanceDiagnosticsController.shared.isEnabled ? DispatchTime.now() : nil
-    }
-
-    private func updateMetrics(_ update: (inout TerminalRenderCoordinatorMetrics) -> Void) {
-        lock.lock()
-        update(&metrics)
-        lock.unlock()
-    }
-
-    private func recordDiagnostics(
-        kind: AlanPerformanceDiagnosticEventKind,
-        durationMs: Double,
-        priority: TerminalRuntimeRenderPriority
-    ) {
-        if let diagnosticsRecorder {
-            guard diagnosticsRecorder.isEnabled else { return }
-        } else {
-            guard AlanPerformanceDiagnosticsController.shared.isEnabled else { return }
-        }
-        let event = AlanPerformanceDiagnosticEvent(
-            kind: kind,
-            durationMs: durationMs,
-            priority: priority.diagnosticsValue,
-            visibility: priority.diagnosticsVisibility,
-            thread: Thread.isMainThread ? "main" : "background"
-        )
-        if let diagnosticsRecorder {
-            diagnosticsRecorder.record(event)
-        } else {
-            AlanPerformanceDiagnosticsController.shared.record(
-                event.kind,
-                durationMs: event.durationMs,
-                paneID: event.paneID,
-                contentID: event.contentID,
-                priority: event.priority,
-                visibility: event.visibility,
-                thread: event.thread,
-                counts: event.counts
-            )
-        }
-    }
-}
-
-func terminalRuntimeRenderPriority(
-    paneID: String,
-    paneSpaceID: String,
-    paneTabID: String,
-    selectedSpaceID: String?,
-    selectedTabID: String?,
-    focusedPaneID: String?,
-    visiblePaneIDs: Set<String>,
-    windowIsVisible: Bool
-) -> TerminalRuntimeRenderPriority {
-    guard windowIsVisible,
-          paneSpaceID == selectedSpaceID,
-          paneTabID == selectedTabID,
-          visiblePaneIDs.contains(paneID)
-    else {
-        return .hiddenBackground
-    }
-
-    guard paneID == focusedPaneID else {
-        return .visibleBackground
-    }
-    return .foregroundInteractive
-}
-
-enum TerminalRuntimePublicationPolicy {
-    static func shouldProjectToShell(
-        previous: TerminalHostRuntimeSnapshot?,
-        next: TerminalHostRuntimeSnapshot
-    ) -> Bool {
-        guard let previous else {
-            return true
-        }
-
-        return shellProjectionChanged(previous: previous, next: next)
-    }
-
-    private static func shellProjectionChanged(
-        previous: TerminalHostRuntimeSnapshot,
-        next: TerminalHostRuntimeSnapshot
-    ) -> Bool {
-        previous.contentID != next.contentID
-            || previous.paneID != next.paneID
-            || previous.tabID != next.tabID
-            || previous.stage != next.stage
-            || previous.renderPriority != next.renderPriority
-            || previous.displayName != next.displayName
-            || previous.displayID != next.displayID
-            || previous.attachedWindowTitle != next.attachedWindowTitle
-            || previous.renderer.failureReason != next.renderer.failureReason
-            || (previous.renderer.phase != .failed && next.renderer.phase == .failed)
-            || previous.paneMetadata.title != next.paneMetadata.title
-            || previous.paneMetadata.workingDirectory != next.paneMetadata.workingDirectory
-            || previous.paneMetadata.summary != next.paneMetadata.summary
-            || previous.paneMetadata.attention != next.paneMetadata.attention
-            || previous.paneMetadata.processExited != next.paneMetadata.processExited
-            || previous.paneMetadata.lastCommandExitCode != next.paneMetadata.lastCommandExitCode
-            || previous.paneMetadata.activeTaskState != next.paneMetadata.activeTaskState
-            || previous.paneMetadata.activity != next.paneMetadata.activity
-            || previous.paneMetadata.clearsActivity != next.paneMetadata.clearsActivity
-            || previous.surfaceState.readiness != next.surfaceState.readiness
-            || previous.surfaceState.rendererHealth != next.surfaceState.rendererHealth
-            || previous.surfaceState.readonly != next.surfaceState.readonly
-            || previous.surfaceState.terminalMode != next.surfaceState.terminalMode
-            || previous.surfaceState.inputReady != next.surfaceState.inputReady
-            || previous.surfaceState.childExited != next.surfaceState.childExited
-    }
-}
-
-enum TerminalHostStage: String, Equatable {
-    case scaffold
-    case viewAttached = "view_attached"
-    case windowAttached = "window_attached"
-    case focused
-}
-
-enum TerminalRendererKind: String, Equatable {
-    case scaffold
-    case ghosttyLive = "ghostty_live"
-}
-
-enum TerminalRendererPhase: String, Equatable {
-    case pending
-    case libraryReady = "library_ready"
-    case appReady = "app_ready"
-    case surfaceReady = "surface_ready"
-    case firstRefresh = "first_refresh"
-    case failed
-}
-
-struct TerminalRendererSnapshot: Equatable {
-    let kind: TerminalRendererKind
-    let phase: TerminalRendererPhase
-    let summary: String
-    let detail: String?
-    let failureReason: String?
-    let recentEvents: [String]
-
-    var phaseLabel: String {
-        phase.rawValue.replacingOccurrences(of: "_", with: " ")
-    }
-
-    static let placeholder = TerminalRendererSnapshot(
-        kind: .scaffold,
-        phase: .pending,
-        summary: "AppKit terminal scaffold is active.",
-        detail: nil,
-        failureReason: nil,
-        recentEvents: []
-    )
-}
-
-struct TerminalPaneMetadataSnapshot: Equatable {
-    let title: String?
-    let workingDirectory: String?
-    let summary: String?
-    let attention: ShellAttentionState
-    let processExited: Bool
-    let lastCommandExitCode: Int?
-    let lastUpdatedAt: Date?
-    let activeTaskState: ShellTabActiveTaskState?
-    let activity: TerminalActivitySnapshot?
-    let clearsActivity: Bool
-
-    init(
-        title: String?,
-        workingDirectory: String?,
-        summary: String?,
-        attention: ShellAttentionState,
-        processExited: Bool,
-        lastCommandExitCode: Int?,
-        lastUpdatedAt: Date?,
-        activeTaskState: ShellTabActiveTaskState? = .inactive,
-        activity: TerminalActivitySnapshot? = nil,
-        clearsActivity: Bool = false
-    ) {
-        self.title = title
-        self.workingDirectory = workingDirectory
-        self.summary = summary
-        self.attention = attention
-        self.processExited = processExited
-        self.lastCommandExitCode = lastCommandExitCode
-        self.lastUpdatedAt = lastUpdatedAt
-        self.activeTaskState = activeTaskState
-        self.activity = activity
-        self.clearsActivity = clearsActivity
-    }
-
-    static let placeholder = TerminalPaneMetadataSnapshot(
-        title: nil,
-        workingDirectory: nil,
-        summary: nil,
-        attention: .idle,
-        processExited: false,
-        lastCommandExitCode: nil,
-        lastUpdatedAt: nil,
-        activeTaskState: .inactive,
-        activity: nil,
-        clearsActivity: false
-    )
-}
-
-struct TerminalHostRuntimeSnapshot: Equatable {
-    let stage: TerminalHostStage
-    let contentID: String?
-    let paneID: String?
-    let tabID: String?
-    let renderPriority: TerminalRuntimeRenderPriority
-    let logicalSize: CGSize
-    let backingSize: CGSize
-    let displayName: String?
-    let displayID: String?
-    let attachedWindowTitle: String?
-    let isFocused: Bool
-    let renderer: TerminalRendererSnapshot
-    let paneMetadata: TerminalPaneMetadataSnapshot
-    let surfaceState: AlanTerminalSurfaceStateSnapshot
-    let lastUpdatedAt: Date
-
-    init(
-        stage: TerminalHostStage,
-        contentID: String?,
-        paneID: String?,
-        tabID: String?,
-        renderPriority: TerminalRuntimeRenderPriority = .foregroundInteractive,
-        logicalSize: CGSize,
-        backingSize: CGSize,
-        displayName: String?,
-        displayID: String?,
-        attachedWindowTitle: String?,
-        isFocused: Bool,
-        renderer: TerminalRendererSnapshot,
-        paneMetadata: TerminalPaneMetadataSnapshot,
-        surfaceState: AlanTerminalSurfaceStateSnapshot,
-        lastUpdatedAt: Date
-    ) {
-        self.stage = stage
-        self.contentID = contentID
-        self.paneID = paneID
-        self.tabID = tabID
-        self.renderPriority = renderPriority
-        self.logicalSize = logicalSize
-        self.backingSize = backingSize
-        self.displayName = displayName
-        self.displayID = displayID
-        self.attachedWindowTitle = attachedWindowTitle
-        self.isFocused = isFocused
-        self.renderer = renderer
-        self.paneMetadata = paneMetadata
-        self.surfaceState = surfaceState
-        self.lastUpdatedAt = lastUpdatedAt
-    }
-
-    var stageLabel: String {
-        stage.rawValue.replacingOccurrences(of: "_", with: " ")
-    }
-
-    /// Equality that ignores the volatile publish timestamps. Two snapshots that
-    /// only differ in `lastUpdatedAt` represent the same observable shell state,
-    /// so callers can suppress redundant `@Published` mutations (and the
-    /// tree-wide SwiftUI invalidation they trigger).
-    func equalsIgnoringTimestamp(_ other: TerminalHostRuntimeSnapshot) -> Bool {
-        stage == other.stage
-            && contentID == other.contentID
-            && paneID == other.paneID
-            && tabID == other.tabID
-            && renderPriority == other.renderPriority
-            && logicalSize == other.logicalSize
-            && backingSize == other.backingSize
-            && displayName == other.displayName
-            && displayID == other.displayID
-            && attachedWindowTitle == other.attachedWindowTitle
-            && isFocused == other.isFocused
-            && renderer == other.renderer
-            && paneMetadata == other.paneMetadata
-            && surfaceState.equalsIgnoringTimestamp(other.surfaceState)
-    }
-
-    static let placeholder = TerminalHostRuntimeSnapshot(
-        stage: .scaffold,
-        contentID: nil,
-        paneID: nil,
-        tabID: nil,
-        renderPriority: .hiddenBackground,
-        logicalSize: .zero,
-        backingSize: .zero,
-        displayName: nil,
-        displayID: nil,
-        attachedWindowTitle: nil,
-        isFocused: false,
-        renderer: .placeholder,
-        paneMetadata: .placeholder,
-        surfaceState: .placeholder,
-        lastUpdatedAt: .now
-    )
-}
 #endif
